Add more category specification :film_frames:
Also fix bug that caused infinite loop instead of terminating if
all articles looped. Previously this was difficult to test, but
with the ability to specify categories, and the fact that many
articles point to Mathematics (which loops), I was able to finally
get a consistent, controlled environment to test this without
creating a whole testing suite of my own.